requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-dotenv>=1.0
PyJWT[crypto]>=2.8
httpx>=0.27
itsdangerous>=2.2
boto3>=1.28